# src/main/java/com/iti/training/exam/core/controller/generated/student/StudentsController.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.iti.training.exam.core.controller.generated.student;

import com.iti.training.exam.model.generated.student.StudentDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.6.0")
@Validated
@Controller
@Tag(name = "Students", description = "the Students API")
public interface StudentsController {

    /**
     * GET /students/count : Count students with filters
     * Count students based on search filters
     *
     * @param firstName  (optional)
     * @param lastName  (optional)
     * @param email  (optional)
     * @param city  (optional)
     * @param graduationYear  (optional)
     * @param gender  (optional)
     * @param minAge  (optional)
     * @param maxAge  (optional)
     * @return Student count retrieved successfully (status code 200)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "countStudents",
        summary = "Count students with filters",
        description = "Count students based on search filters",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student count retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/students/count",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _countStudents(
        @Parameter(name = "firstName", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "firstName", required = false) String firstName,
        @Parameter(name = "lastName", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "lastName", required = false) String lastName,
        @Parameter(name = "email", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = false) String email,
        @Parameter(name = "city", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "city", required = false) String city,
        @Parameter(name = "graduationYear", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "graduationYear", required = false) Integer graduationYear,
        @Parameter(name = "gender", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "gender", required = false) String gender,
        @Parameter(name = "minAge", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "minAge", required = false) Integer minAge,
        @Parameter(name = "maxAge", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "maxAge", required = false) Integer maxAge
    );


    /**
     * POST /students : Create a new student
     * Create a new student record
     *
     * @param studentDTO  (required)
     * @return Student created successfully (status code 200)
     *         or Bad request (status code 400)
     *         or Student already exists (status code 409)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "createStudent",
        summary = "Create a new student",
        description = "Create a new student record",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request"),
            @ApiResponse(responseCode = "409", description = "Student already exists"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/students",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _createStudent(
        @Parameter(name = "StudentDTO", description = "", required = true) @Valid @RequestBody StudentDTO studentDTO
    );


    /**
     * DELETE /students/{ssn} : Delete student
     * Delete a student record by SSN
     *
     * @param ssn  (required)
     * @return Student deleted successfully (status code 200)
     *         or Student not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "deleteStudent",
        summary = "Delete student",
        description = "Delete a student record by SSN",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student deleted successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Student not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/students/{ssn}",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _deleteStudent(
        @Pattern(regexp = "^\\d{14}$") @Parameter(name = "ssn", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ssn") String ssn
    );


    /**
     * GET /students/{ssn}/exists : Check if student exists
     * Check if a student exists by SSN
     *
     * @param ssn  (required)
     * @return Existence check completed (status code 200)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "existsBySsn",
        summary = "Check if student exists",
        description = "Check if a student exists by SSN",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Existence check completed", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/students/{ssn}/exists",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _existsBySsn(
        @Pattern(regexp = "^\\d{14}$") @Parameter(name = "ssn", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ssn") String ssn
    );


    /**
     * GET /students : Get students with filters
     * Retrieve students based on search filters
     *
     * @param firstName  (optional)
     * @param lastName  (optional)
     * @param email  (optional)
     * @param city  (optional)
     * @param graduationYear  (optional)
     * @param gender  (optional)
     * @param minAge  (optional)
     * @param maxAge  (optional)
     * @param pageNum Pagination Page Number (optional, default to 0)
     * @param pageSize Pagination Page Size (optional, default to 25)
     * @param noPagination If true, returns all results without pagination. Useful for dropdowns or exports.  (optional, default to false)
     * @param sortBy Sort by field (optional)
     * @param sortDir Sort direction (optional)
     * @return Students retrieved successfully (status code 200)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getAllByFilters",
        summary = "Get students with filters",
        description = "Retrieve students based on search filters",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Students retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/students",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _getAllByFilters(
        @Parameter(name = "firstName", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "firstName", required = false) String firstName,
        @Parameter(name = "lastName", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "lastName", required = false) String lastName,
        @Parameter(name = "email", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = false) String email,
        @Parameter(name = "city", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "city", required = false) String city,
        @Parameter(name = "graduationYear", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "graduationYear", required = false) Integer graduationYear,
        @Parameter(name = "gender", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "gender", required = false) String gender,
        @Parameter(name = "minAge", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "minAge", required = false) Integer minAge,
        @Parameter(name = "maxAge", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "maxAge", required = false) Integer maxAge,
        @Parameter(name = "pageNum", description = "Pagination Page Number", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageNum", required = false, defaultValue = "0") Integer pageNum,
        @Min(10) @Max(100) @Parameter(name = "pageSize", description = "Pagination Page Size", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "25") Integer pageSize,
        @Parameter(name = "noPagination", description = "If true, returns all results without pagination. Useful for dropdowns or exports. ", in = ParameterIn.QUERY) @Valid @RequestParam(value = "noPagination", required = false, defaultValue = "false") Boolean noPagination,
        @Parameter(name = "sortBy", description = "Sort by field", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortBy", required = false) String sortBy,
        @Parameter(name = "sortDir", description = "Sort direction", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortDir", required = false) String sortDir
    );


    /**
     * GET /students/progress : Get all students progress with filters
     * Retrieve progress information for all students based on filters
     *
     * @param firstName  (optional)
     * @param lastName  (optional)
     * @param email  (optional)
     * @param city  (optional)
     * @param graduationYear  (optional)
     * @param gender  (optional)
     * @param minAge  (optional)
     * @param maxAge  (optional)
     * @param pageNum Pagination Page Number (optional, default to 0)
     * @param pageSize Pagination Page Size (optional, default to 25)
     * @param noPagination If true, returns all results without pagination. Useful for dropdowns or exports.  (optional, default to false)
     * @return Students progress retrieved successfully (status code 200)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getAllStudentsProgressByFilter",
        summary = "Get all students progress with filters",
        description = "Retrieve progress information for all students based on filters",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Students progress retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/students/progress",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _getAllStudentsProgressByFilter(
        @Parameter(name = "firstName", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "firstName", required = false) String firstName,
        @Parameter(name = "lastName", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "lastName", required = false) String lastName,
        @Parameter(name = "email", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = false) String email,
        @Parameter(name = "city", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "city", required = false) String city,
        @Parameter(name = "graduationYear", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "graduationYear", required = false) Integer graduationYear,
        @Parameter(name = "gender", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "gender", required = false) String gender,
        @Parameter(name = "minAge", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "minAge", required = false) Integer minAge,
        @Parameter(name = "maxAge", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "maxAge", required = false) Integer maxAge,
        @Parameter(name = "pageNum", description = "Pagination Page Number", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageNum", required = false, defaultValue = "0") Integer pageNum,
        @Min(10) @Max(100) @Parameter(name = "pageSize", description = "Pagination Page Size", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "25") Integer pageSize,
        @Parameter(name = "noPagination", description = "If true, returns all results without pagination. Useful for dropdowns or exports. ", in = ParameterIn.QUERY) @Valid @RequestParam(value = "noPagination", required = false, defaultValue = "false") Boolean noPagination
    );


    /**
     * GET /students/{ssn} : Get student by SSN
     * Retrieve a specific student by their SSN
     *
     * @param ssn  (required)
     * @return Student retrieved successfully (status code 200)
     *         or Student not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getStudentBySsn",
        summary = "Get student by SSN",
        description = "Retrieve a specific student by their SSN",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Student not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/students/{ssn}",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _getStudentBySsn(
        @Pattern(regexp = "^\\d{14}$") @Parameter(name = "ssn", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ssn") String ssn
    );


    /**
     * GET /students/{ssn}/exam-history : Get student exam history
     * Retrieve exam attempt history for a specific student
     *
     * @param ssn  (required)
     * @return Exam history retrieved successfully (status code 200)
     *         or Student not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getStudentExamHistory",
        summary = "Get student exam history",
        description = "Retrieve exam attempt history for a specific student",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Exam history retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Student not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/students/{ssn}/exam-history",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _getStudentExamHistory(
        @Pattern(regexp = "^\\d{14}$") @Parameter(name = "ssn", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ssn") String ssn
    );


    /**
     * GET /students/{ssn}/progress : Get student progress
     * Retrieve progress information for a specific student
     *
     * @param ssn  (required)
     * @return Student progress retrieved successfully (status code 200)
     *         or Student not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getStudentProgress",
        summary = "Get student progress",
        description = "Retrieve progress information for a specific student",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student progress retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Student not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/students/{ssn}/progress",
        produces = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _getStudentProgress(
        @Pattern(regexp = "^\\d{14}$") @Parameter(name = "ssn", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ssn") String ssn
    );


    /**
     * PUT /students/{ssn} : Update student
     * Update an existing student record
     *
     * @param ssn  (required)
     * @param studentDTO  (required)
     * @return Student updated successfully (status code 200)
     *         or Student not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "updateStudent",
        summary = "Update student",
        description = "Update an existing student record",
        tags = { "Students" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Student updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = com.iti.training.exam.core.controller.config.ApiResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Student not found"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/students/{ssn}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<com.iti.training.exam.core.controller.config.ApiResponse> _updateStudent(
        @Pattern(regexp = "^\\d{14}$") @Parameter(name = "ssn", description = "", required = true, in = ParameterIn.PATH) @PathVariable("ssn") String ssn,
        @Parameter(name = "StudentDTO", description = "", required = true) @Valid @RequestBody StudentDTO studentDTO
    );

}